requirements.txt
amqp==2.1.4
asn1crypto==0.22.0
billiard==3.5.0.2
boto3==1.4.4
botocore==1.5.60
celery==4.0.2
certifi==2017.4.17
cffi==1.10.0
chardet==3.0.3
cryptography==1.9
defusedxml==0.5.0
dj-database-url==0.4.2
dj-static==0.0.6
Django==1.11.2
django-admin-tools==0.8.1
django-admin-tools-stats==0.8.0
django-allauth==0.32.0
django-appconf==1.0.2
django-baton==1.0.1
django-bower==5.2.0
django-braces==1.11.0
django-cache-utils==0.7.2
django-classy-tags==0.8.0
django-crispy-forms==1.6.1
django-easy-maps==0.9.3
django-filter==1.0.4
django-generic-positions==0.2.2
django-grappelli==2.10.1
django-hvad==1.8.0
django-imagekit==4.0.1
django-jet==1.0.6
django-jsonfield==1.0.1
django-libs==1.67.12
django-model-utils==3.0.0
django-nvd3==0.9.7
django-private-chat==0.1.6
django-qsstats-magic==0.7.2
django-review==1.9.6
django-star-ratings==0.5.5
django-storages==1.5.2
django-toolbelt==0.0.1
django-user-media==1.2.4
django-viewflow==1.0.1
docutils==0.13.1
easy-thumbnails==2.4.1
geopy==1.11.0
google-api-python-client==1.6.2
gunicorn==19.7.1
httplib2==0.10.3
idna==2.5
Jinja2==2.9.6
jmespath==0.9.3
kombu==4.0.2
MarkupSafe==1.0
oauth2client==4.1.0
oauthlib==2.0.2
olefile==0.44
pep8==1.7.0
pilkit==2.0
Pillow==4.1.1
pinax-notifications==4.0.0
psycopg2==2.7.1
pyasn1==0.2.3
pyasn1-modules==0.0.9
pycparser==2.17
pyOpenSSL==17.0.0
python-dateutil==2.6.0
python-decouple==3.0
python-nvd3==0.14.2
python-slugify==1.1.4
python3-openid==3.1.0
pytz==2017.2
redis==2.10.5
requests==2.16.5
requests-oauthlib==0.8.0
rsa==3.4.2
s3transfer==0.1.10
simplejson==3.10.0
six==1.10.0
static3==0.7.0
Unidecode==0.4.20
uritemplate==3.0.0
urllib3==1.21.1
vine==1.1.3
websockets==3.2
